+++ src/game.hpp
@@ -23,7 +23,7 @@ class Game {
   int GetLevel();
 
   /* Operations for money. */
-  const std::uint32_t &PlayerMoney() const;
+  const std::uint64_t &PlayerMoney() const;
   void AddMoney(std::uint32_t money) { playerMoney_ += money; }
   void RemoveMoney(std::uint32_t money) { playerMoney_ -= money <= playerMoney_ ? money : playerMoney_; }
 
@@ -32,6 +32,6 @@ class Game {
   int level_;
   GameGrid grid_;
   //int enemyMoney_;
-  std::uint32_t playerMoney_;
+  std::uint64_t playerMoney_;
   std::vector<GameObject*> objects_;
 };
